:truck: rename: view 파일들 디렉토리 분리
